move interp APIs to Project
